Week7 (31-35)/Day34/rag_deploy_demo/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.7,<3
python-dotenv==1.0.1
openai==1.45.0
httpx==0.27.2
pytest==8.3.3
